modularized mvp functions
