+++ src/index.html
@@ -7,7 +7,7 @@
 </head>
 <body>
 
-  <div id="cardHolder"></div>
+  <div class="card-holder" id="cardHolder"></div>
   <div id="bearForm"></div>
 
 </body>
